experimental/ThaiS2T/requirements.txt
virtualenv
google-cloud
google-cloud-speech
ffmpeg-python
numpy
matplotlib
scipy
sklearn
hmmlearn
simplejson
eyed3
pydub
pysrt
pythainlp


